jartool.c (extract_jar): Changed rdamt to size_t. pushback.c (pb_push): Changed in_amt, wrap to size_t. Added a size_t cast, and removed an int cast. (pb_read) Changed out_amt, wrap, tmp to size_t. Added a size_t cast, and removed an int cast. Added new variable num_read. pushback.h (pb_push, pb_read): Changed prototypes to use size_t.

+++ jartool.c
@@ -1513,7 +1513,7 @@ int create_central_header(int fd){
 }
 
 int extract_jar(int fd, const char **files, int file_num){
-  int rdamt;
+  size_t rdamt;
   int out_a, in_a;
   ub4 signature;
   ub4 csize;
